fix(graphql): rate limit on parse error (#998)

+++ config/graphql.php
@@ -6,7 +6,7 @@ use App\GraphQL\Schema\Schemas\DefaultSchema;
 use App\GraphQL\Schema\Types\Base\PaginationInfoType;
 
 return [
-    'rate_limit' => env('GRAPHQL_RATE_LIMIT', 80),
+    'rate_limit' => (int) env('GRAPHQL_RATE_LIMIT', 80),
 
     'route' => [
         // The prefix for routes; do NOT use a leading slash!
